requirements.txt
openpyxl==3.1.2
psycopg2-binary==2.9.5
loguru==0.6.0
PyQt5==5.15.9
PyInstaller==5.12.0
